# Conflicts:
#	code/__DEFINES/hijack.dm
#	code/controllers/subsystem/hijack.dm
#	colonialmarines.dme
#	tgui/packages/tgui/interfaces/EscapePodConsole.tsx

+++ code/game/turfs/walls/walls.dm
@@ -5,8 +5,8 @@
 	icon_state = "0"
 	opacity = TRUE
 	layer = WALL_LAYER
+	is_weedable = FULLY_WEEDABLE
 	/// 1 = Can't be deconstructed by tools or thermite. Used for Sulaco walls
-	var/hull = 0
 	var/walltype = WALL_METAL
 	/// when walls smooth with one another, the type of junction each wall is.
 	var/junctiontype
@@ -47,8 +47,15 @@
 	var/list/blend_objects = list(/obj/structure/machinery/door, /obj/structure/window_frame, /obj/structure/window/framed) // Objects which to blend with
 	var/list/noblend_objects = list(/obj/structure/machinery/door/window) //Objects to avoid blending with (such as children of listed blend objects.
 
+	var/list/hiding_humans = list()
+
+	///Whether this turf is currently being manipulated to prevent doubling up
+	var/busy = FALSE
+
 /turf/closed/wall/Initialize(mapload, ...)
 	. = ..()
+	is_weedable = initial(is_weedable) //so we can spawn weeds on the wall
+
 	// Defer updating based on neighbors while we're still loading map
 	if(mapload && . != INITIALIZE_HINT_QDEL)
 		return INITIALIZE_HINT_LATELOAD
@@ -56,6 +63,7 @@
 	update_connections(TRUE)
 	if(. != INITIALIZE_HINT_LATELOAD)
 		update_icon()
+
 
 /turf/closed/wall/LateInitialize()
 	. = ..()
@@ -103,7 +111,7 @@
 		var/list/turf/cardinal_neighbors = list(get_step(src, NORTH), get_step(src, SOUTH), get_step(src, EAST), get_step(src, WEST))
 		for(var/turf/cardinal_turf as anything in cardinal_neighbors)
 			for(var/obj/structure/bed/nest/found_nest in cardinal_turf)
-				if(found_nest.dir == get_dir(found_nest, src))
+				if(found_nest.dir == get_dir(found_nest, src) && !density)
 					qdel(found_nest) //nests are built on walls, no walls, no nest
 
 /turf/closed/wall/MouseDrop_T(mob/current_mob, mob/user)
@@ -118,14 +126,92 @@
 	if(isxeno(user) && istype(user.get_active_hand(), /obj/item/grab))
 		var/mob/living/carbon/xenomorph/user_as_xenomorph = user
 		user_as_xenomorph.do_nesting_host(current_mob, src)
+
+	// wall leaning by androbetel
+	if(!ishuman(current_mob))
+		return
+
+	if(current_mob != user)
+		return
+	var/mob/living/carbon/hiding_human = current_mob
+	var/can_lean = TRUE
+
+	if(istype(user.l_hand, /obj/item/grab) || istype(user.r_hand, /obj/item/grab))
+		to_chat(user, SPAN_WARNING("You can't lean while grabbing someone!"))
+		can_lean = FALSE
+	if(current_mob.is_mob_incapacitated())
+		to_chat(user, SPAN_WARNING("You can't lean while incapacitated!"))
+		can_lean = FALSE
+	if(current_mob.resting)
+		to_chat(user, SPAN_WARNING("You can't lean while resting!"))
+		can_lean = FALSE
+	if(current_mob.buckled)
+		to_chat(user, SPAN_WARNING("You can't lean while buckled!"))
+		can_lean = FALSE
+
+	var/direction = get_dir(src, current_mob)
+	var/shift_pixel_x = 0
+	var/shift_pixel_y = 0
+
+	if(!can_lean)
+		return
+	switch(direction)
+		if(NORTH)
+			shift_pixel_y = -10
+		if(SOUTH)
+			shift_pixel_y = 16
+		if(WEST)
+			shift_pixel_x = 10
+		if(EAST)
+			shift_pixel_x = -10
+		else
+			return
+
+	for(var/mob/living/carbon/human/hiding in hiding_humans)
+		if(hiding_humans[hiding] == direction)
+			return
+
+	hiding_humans += current_mob
+	hiding_humans[current_mob] = direction
+	hiding_human.Moved() //just to be safe
+	hiding_human.setDir(direction)
+	animate(hiding_human, pixel_x = shift_pixel_x, pixel_y = shift_pixel_y, time = 1)
+	if(direction == NORTH)
+		hiding_human.add_filter("cutout", 1, alpha_mask_filter(icon = icon('icons/effects/effects.dmi', "cutout")))
+	ADD_TRAIT(hiding_human, TRAIT_UNDENSE, WALL_HIDING_TRAIT)
+	RegisterSignal(hiding_human, list(COMSIG_MOVABLE_MOVED, COMSIG_LIVING_SET_BODY_POSITION, COMSIG_MOB_RESISTED, COMSIG_MOB_ANIMATING), PROC_REF(unhide_human), hiding_human)
 	..()
+
+/turf/closed/wall/proc/unhide_human(mob/living/carbon/human/to_unhide)
+	SIGNAL_HANDLER
+	if(!to_unhide)
+		return
+
+	REMOVE_TRAIT(to_unhide, TRAIT_UNDENSE, WALL_HIDING_TRAIT)
+	to_unhide.pixel_x = initial(to_unhide.pixel_x)
+	to_unhide.pixel_y = initial(to_unhide.pixel_y)
+	to_unhide.layer = initial(to_unhide.layer)
+	to_unhide.apply_effect(1, SUPERSLOW)
+	to_unhide.apply_effect(2, SLOW)
+	hiding_humans -= to_unhide
+	UnregisterSignal(to_unhide, list(COMSIG_MOVABLE_MOVED, COMSIG_LIVING_SET_BODY_POSITION, COMSIG_MOB_RESISTED, COMSIG_MOB_ANIMATING))
+	to_chat(to_unhide, SPAN_WARNING("You couldn't sit still so you stop leaning on the wall!"))
+	to_unhide.remove_filter("cutout")
+
+/turf/closed/wall/Destroy()
+	if(hiding_humans.len)
+		for(var/mob/living/carbon/human/human in hiding_humans)
+			unhide_human(human)
+
+	return ..()
+
 
 /turf/closed/wall/attack_alien(mob/living/carbon/xenomorph/user)
 	if(acided_hole && user.mob_size >= MOB_SIZE_BIG)
 		acided_hole.expand_hole(user) //This proc applies the attack delay itself.
 		return XENO_NO_DELAY_ACTION
 
-	if(!hull && user.claw_type >= claws_minimum && !acided_hole)
+	if(!(turf_flags & TURF_HULL) && user.claw_type >= claws_minimum && !acided_hole && user.a_intent == INTENT_HARM)
 		user.animation_attack_on(src)
 		playsound(src, 'sound/effects/metalhit.ogg', 25, 1)
 		if(damage >= (damage_cap - (damage_cap / XENO_HITS_TO_DESTROY_WALL)))
@@ -158,7 +244,7 @@
 /turf/closed/wall/get_examine_text(mob/user)
 	. = ..()
 
-	if(hull)
+	if(turf_flags & TURF_HULL)
 		.+= SPAN_WARNING("You don't think you have any tools able to even scratch this.")
 		return //If it's indestructable, we don't want to give the wrong impression by saying "you can decon it with a welder"
 
@@ -196,7 +282,7 @@
 
 //Damage
 /turf/closed/wall/proc/take_damage(dam, mob/M)
-	if(hull) //Hull is literally invincible
+	if(turf_flags & TURF_HULL) //Hull is literally invincible
 		return
 	if(!dam)
 		return
@@ -230,7 +316,7 @@
 // Walls no longer spawn a metal sheet when destroyed to reduce clutter and
 // improve visual readability.
 /turf/closed/wall/proc/dismantle_wall(devastated = 0, explode = 0)
-	if(hull) //Hull is literally invincible
+	if(turf_flags & TURF_HULL) //Hull is literally invincible
 		return
 	if(devastated)
 		make_girder(TRUE)
@@ -242,7 +328,7 @@
 	ScrapeAway()
 
 /turf/closed/wall/ex_act(severity, explosion_direction, datum/cause_data/cause_data)
-	if(hull)
+	if(turf_flags & TURF_HULL)
 		return
 	var/location = get_step(get_turf(src), explosion_direction) // shrapnel will just collide with the wall otherwise
 	var/exp_damage = severity*EXPLOSION_DAMAGE_MULTIPLIER_WALL
@@ -256,6 +342,8 @@
 		dismantle_wall(FALSE, TRUE)
 		if(!istype(src, /turf/closed/wall/resin))
 			create_shrapnel(location, rand(2,5), explosion_direction, , /datum/ammo/bullet/shrapnel/light, cause_data)
+		else
+			create_shrapnel(location, rand(2,5), explosion_direction, , /datum/ammo/bullet/shrapnel/light/resin, cause_data)
 	else
 		if(istype(src, /turf/closed/wall/resin))
 			exp_damage *= RESIN_EXPLOSIVE_MULTIPLIER
@@ -268,7 +356,7 @@
 	return
 
 /turf/closed/wall/get_explosion_resistance()
-	if(hull)
+	if(turf_flags & TURF_HULL)
 		return 1000000
 
 	return (damage_cap - damage)/EXPLOSION_DAMAGE_MULTIPLIER_WALL
@@ -277,7 +365,7 @@
 	if(melting)
 		to_chat(user, SPAN_WARNING("The wall is already burning with thermite!"))
 		return
-	if(hull)
+	if(turf_flags & TURF_HULL)
 		return
 	melting = TRUE
 
@@ -322,7 +410,7 @@
 //Interactions
 /turf/closed/wall/attack_animal(mob/living/M as mob)
 	if(M.wall_smash)
-		if((istype(src, /turf/closed/wall/r_wall)) || hull)
+		if((istype(src, /turf/closed/wall/r_wall)) || turf_flags & TURF_HULL)
 			to_chat(M, SPAN_WARNING("This [name] is far too strong for you to destroy."))
 			return
 		else
@@ -343,15 +431,20 @@
 		var/obj/item/grab/attacker_grab = attacking_item
 		var/mob/living/carbon/xenomorph/user_as_xenomorph = user
 		user_as_xenomorph.do_nesting_host(attacker_grab.grabbed_thing, src)
+		return
 
 	if(!ishuman(user))
 		to_chat(user, SPAN_WARNING("You don't have the dexterity to do this!"))
 		return
 
+	if(busy)
+		to_chat(user, SPAN_WARNING("Someone else is already working on [src]."))
+		return
+
 	//THERMITE related stuff. Calls src.thermitemelt() which handles melting simulated walls and the relevant effects
 	if(thermite)
 		if(attacking_item.heat_source >= 1000)
-			if(hull)
+			if(turf_flags & TURF_HULL)
 				to_chat(user, SPAN_WARNING("[src] is much too tough for you to do anything to it with [attacking_item]."))
 			else
 				if(iswelder(attacking_item))
@@ -365,17 +458,20 @@
 		if(user.action_busy)
 			return
 		if(!(HAS_TRAIT(user, TRAIT_SUPER_STRONG) || !current_hammer.really_heavy))
-			to_chat(user, SPAN_WARNING("You can't use \the [current_hammer] properly!"))
+			to_chat(user, SPAN_WARNING("You can't use [current_hammer] properly!"))
 			return
-		if(hull)
-			to_chat(user, SPAN_WARNING("Even with your immense strength, you can't bring down \the [src]."))
+		if(turf_flags & TURF_HULL)
+			to_chat(user, SPAN_WARNING("Even with your immense strength, you can't bring down [src]."))
 			return
 
-		to_chat(user, SPAN_NOTICE("You start taking down \the [src]."))
+		to_chat(user, SPAN_NOTICE("You start taking down [src]."))
+		busy = TRUE
 		if(!do_after(user, 5 SECONDS, INTERRUPT_ALL_OUT_OF_RANGE, BUSY_ICON_BUILD))
-			to_chat(user, SPAN_NOTICE("You stop taking down \the [src]."))
+			busy = FALSE
+			to_chat(user, SPAN_NOTICE("You stop taking down [src]."))
 			return
-		to_chat(user, SPAN_NOTICE("You tear down \the [src]."))
+		busy = FALSE
+		to_chat(user, SPAN_NOTICE("You tear down [src]."))
 
 		playsound(src, 'sound/effects/meteorimpact.ogg', 40, 1)
 		playsound(src, 'sound/effects/ceramic_shatter.ogg', 40, 1)
@@ -417,8 +513,9 @@
 		to_chat(user, SPAN_NOTICE("You place the torch down on the wall."))
 		new /obj/structure/prop/brazier/frame/full/torch(src)
 		qdel(attacking_item)
+		return
 
-	if(hull)
+	if(turf_flags & TURF_HULL)
 		to_chat(user, SPAN_WARNING("[src] is much too tough for you to do anything to it with [attacking_item]."))
 		return
 
@@ -443,8 +540,11 @@
 				user.visible_message(SPAN_NOTICE("[user] begins removing the support lines."),
 				SPAN_NOTICE("You begin removing the support lines."))
 				playsound(src, 'sound/items/Screwdriver.ogg', 25, 1)
+				busy = TRUE
 				if(!do_after(user, 60 * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION), INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+					busy = FALSE
 					return
+				busy = FALSE
 				d_state = WALL_STATE_WIRECUTTER
 				user.visible_message(SPAN_NOTICE("[user] removes the support lines."), SPAN_NOTICE("You remove the support lines."))
 				return
@@ -454,8 +554,11 @@
 				user.visible_message(SPAN_NOTICE("[user] begins uncrimping the hydraulic lines."),
 				SPAN_NOTICE("You begin uncrimping the hydraulic lines."))
 				playsound(src, 'sound/items/Wirecutter.ogg', 25, 1)
+				busy = TRUE
 				if(!do_after(user, 60 * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION), INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+					busy = FALSE
 					return
+				busy = FALSE
 				d_state = WALL_STATE_WRENCH
 				user.visible_message(SPAN_NOTICE("[user] finishes uncrimping the hydraulic lines."), SPAN_NOTICE("You finish uncrimping the hydraulic lines."))
 				return
@@ -465,8 +568,11 @@
 				user.visible_message(SPAN_NOTICE("[user] starts loosening the anchoring bolts securing the support rods."),
 				SPAN_NOTICE("You start loosening the anchoring bolts securing the support rods."))
 				playsound(src, 'sound/items/Ratchet.ogg', 25, 1)
+				busy = TRUE
 				if(!do_after(user, 60 * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION), INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+					busy = FALSE
 					return
+				busy = FALSE
 				d_state = WALL_STATE_CROWBAR
 				user.visible_message(SPAN_NOTICE("[user] removes the bolts anchoring the support rods."), SPAN_NOTICE("You remove the bolts anchoring the support rods."))
 				return
@@ -476,7 +582,12 @@
 				user.visible_message(SPAN_NOTICE("[user] struggles to pry apart the connecting rods."),
 				SPAN_NOTICE("You struggle to pry apart the connecting rods."))
 				playsound(src, 'sound/items/Crowbar.ogg', 25, 1)
+				busy = TRUE
 				if(!do_after(user, 60 * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION), INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+					busy = FALSE
+					return
+				busy = FALSE
+				if(!istype(src, /turf/closed/wall))
 					return
 				user.visible_message(SPAN_NOTICE("[user] pries apart the connecting rods."), SPAN_NOTICE("You pry apart the connecting rods."))
 				new /obj/item/stack/rods(src)
@@ -490,16 +601,21 @@
 		return FALSE
 	if(user.a_intent != INTENT_HELP)
 		return FALSE
+	if(busy)
+		return FALSE
 
 	var/obj/item/tool/weldingtool/WT = W
 	if(WT.remove_fuel(0, user))
 		user.visible_message(SPAN_NOTICE("[user] starts repairing the damage to [src]."),
 		SPAN_NOTICE("You start repairing the damage to [src]."))
 		playsound(src, 'sound/items/Welder.ogg', 25, 1)
+		busy = TRUE
 		if(do_after(user, max(5, floor(damage / 5) * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION)), INTERRUPT_ALL, BUSY_ICON_FRIENDLY) && istype(src, /turf/closed/wall) && WT && WT.isOn())
+			busy = FALSE
 			user.visible_message(SPAN_NOTICE("[user] finishes repairing the damage to [src]."),
 			SPAN_NOTICE("You finish repairing the damage to [src]."))
 			take_damage(-damage)
+		busy = FALSE
 	else
 		to_chat(user, SPAN_WARNING("You need more welding fuel to complete this task."))
 
@@ -514,14 +630,19 @@
 		return
 	if(user.a_intent == INTENT_HELP)
 		return
+	if(busy)
+		return
 
 	playsound(src, 'sound/items/Welder.ogg', 25, 1)
 	user.visible_message(SPAN_NOTICE("[user] begins slicing through the outer plating."),
 	SPAN_NOTICE("You begin slicing through the outer plating."))
 	if(!WT || !WT.isOn())
 		return
+	busy = TRUE
 	if(!do_after(user, 60 * user.get_skill_duration_multiplier(SKILL_CONSTRUCTION), INTERRUPT_ALL|BEHAVIOR_IMMOBILE, BUSY_ICON_BUILD))
+		busy = FALSE
 		return
+	busy = FALSE
 	d_state = WALL_STATE_SCREW
 	user.visible_message(SPAN_NOTICE("[user] slices through the outer plating."), SPAN_NOTICE("You slice through the outer plating."))
 	return
@@ -597,4 +718,4 @@
 	return TRUE
 
 /turf/closed/wall/can_be_dissolved()
-	return !hull
+	return !(turf_flags & TURF_HULL)
